notesharing-notes/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
boto3==1.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.1
groq==0.11.0
aiofiles==24.1.0
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
